+++ tests/integration_tests/test_priority.py
@@ -120,29 +120,29 @@ def test_native_tx_priority(ethermint: Ethermint):
         {
             "from": eth_to_bech32(ADDRS["community"]),
             "to": eth_to_bech32(ADDRS["validator"]),
-            "amount": "1000aphoton",
-            "gas_prices": f"{base_fee + PRIORITY_REDUCTION * 600000}aphoton",
+            "amount": "1000aENTGL",
+            "gas_prices": f"{base_fee + PRIORITY_REDUCTION * 600000}aENTGL",
             "max_priority_price": 0,
         },
         {
             "from": eth_to_bech32(ADDRS["signer1"]),
             "to": eth_to_bech32(ADDRS["signer2"]),
-            "amount": "1000aphoton",
-            "gas_prices": f"{base_fee + PRIORITY_REDUCTION * 600000}aphoton",
+            "amount": "1000aENTGL",
+            "gas_prices": f"{base_fee + PRIORITY_REDUCTION * 600000}aENTGL",
             "max_priority_price": PRIORITY_REDUCTION * 200000,
         },
         {
             "from": eth_to_bech32(ADDRS["signer2"]),
             "to": eth_to_bech32(ADDRS["signer1"]),
-            "amount": "1000aphoton",
-            "gas_prices": f"{base_fee + PRIORITY_REDUCTION * 400000}aphoton",
+            "amount": "1000aENTGL",
+            "gas_prices": f"{base_fee + PRIORITY_REDUCTION * 400000}aENTGL",
             "max_priority_price": PRIORITY_REDUCTION * 400000,
         },
         {
             "from": eth_to_bech32(ADDRS["validator"]),
             "to": eth_to_bech32(ADDRS["community"]),
-            "amount": "1000aphoton",
-            "gas_prices": f"{base_fee + PRIORITY_REDUCTION * 600000}aphoton",
+            "amount": "1000aENTGL",
+            "gas_prices": f"{base_fee + PRIORITY_REDUCTION * 600000}aENTGL",
             "max_priority_price": None,  # no extension, maximum tipFeeCap
         },
     ]
@@ -161,7 +161,7 @@ def test_native_tx_priority(ethermint: Ethermint):
                 tx, tc["from"], max_priority_price=tc.get("max_priority_price")
             )
         )
-        gas_price = int(tc["gas_prices"].removesuffix("aphoton"))
+        gas_price = int(tc["gas_prices"].removesuffix("aENTGL"))
         expect_priorities.append(
             min(
                 get_max_priority_price(tc.get("max_priority_price")),